refactor(content): decompose liferay-content-prune into focused internal modules
Split the 987-line monolith into four internal modules plus a thin facade:
- liferay-content-prune-context.ts: validation, context factory, gateway error helpers
- liferay-content-prune-folders.ts: folder tree discovery, counting, removable folder computation
- liferay-content-prune-plan.ts: article fetching, plan building, summary, dedup, filtering, sample articles
- liferay-content-prune-execute.ts: deletion execution (whole-folder and per-article)
- liferay-content-prune.ts: thin facade with public types and orchestration only

Fix post-decomposition review findings:
- F1: break circular import by moving StructurePruneSummary and SampleArticle into plan module
- F2: remove unused collectFolderTree import from execute module
- F3: move collectPruneArticles and buildPrunePlanSummary into plan module; decouple
  buildPrunePlanSummary signature from ContentPruneOptions to avoid re-introducing the circular dep;
  remove ARTICLE_INVENTORY_CONCURRENCY, runConcurrent, computeRemovableFolderIds, Printer,
  LiferayGateway, and ArticleWithFolder from the facade
